01-opentelemetry/app/requirements.txt
opentelemetry-sdk
opentelemetry-exporter-prometheus
Flask
